Added first version of async-await pattern
